- reactivated the preconditioner option SymmetricGaussSeidel and Jacobi that were previously broken. Now using Ifpack implementation of such methods.
- reactivated the AZOMEGA parameter that was previously unused/broken to be the damping factor for Jacobi/GaussSeidel type methods.

- added additional preconditioner options GaussSeidel and DownwindGaussSeidel (Ifpack implementations will be used)

- added additional smoother options GS and DGS (meaning the same as above) to the ML multigrid block

The intention is to create a lightweight implicit reindexing in downwind direction following papers by Bey and Wittum as well as Hackbusch.
This is work in progress but the hookup in baci is already there.
(Currently choosing DownwindGaussSeidel or DGS will default to standard GaussSeidel and GS without downwinding)

Once downwinding is present, I expect much more efficient MG smoothers for Navier Stokes problems. I also plan to use it in the SIMPLE approaches where one could spare the setup of an AMG on the velocity dofs (replace with DownwindGaussSeidel) and just do AMG on the pressure equations. It turned out that the momentum equation alone is so easy to solve that doing AMG there is overkill and one never amortizes the setup cost.

This submit was presented to you by the FEIBI initiative (excessively trying to keep the ratio element time / solver time larger than one ;-) )

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5328 060e22d8-9d1c-0410-868a-e3a950d63d95