improved documentation
Note: the keyword \class is not required, if the Doxygen documentation is directly above the class definition in the header file.
So far, most classes were documented twice - at the beginning of the file and at the class definitions
Doxygen was complaining and using any documentation that comes first. Add to that the copy and paste culture and we have one class documented many times with different text - adding funny randomness to the already complex documentation effort.

-> removed all \class keywords

-> added proper \file keywords in may files

more folders to come in a minute


this commit was presented and supported by SOCKE:

                     -~-
                     |~|
                     | |_
                     |SOCKE

-> Systematic Online-Code-Knowledge Enhancement <-

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5423 060e22d8-9d1c-0410-868a-e3a950d63d95