fix monolithic FSI with generalized alpha fluid
Given this framework it is now possible to solve FSI problems monolithically
using the same time discretization scheme for both fluid and structure field.
This way more that just backward Euler is available.

Note: Right now both fluid and structure input contains alpha_f, alpha_m and
gamma factors. You must specify matching values to do a meaningful calculation.
And the structure uses a different convention for alpha_f and alpha_m than
the fluid. (Maybe we can unify things here.)

This is the first version that is supposed to work. Clean- and speedup is needed.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5447 060e22d8-9d1c-0410-868a-e3a950d63d95