implemented check for ZIG-ZAGGING of the active contact set
A problem of the active set strategy which sometimes arises is known
from optimization literature as jamming or zig-zagging. This means
that within a load/time-step the algorithm can have more than one
solution due to the fact that the active set is not unique. Hence the
algorithm jumps between the solutions of the active set.

To overcome this problem we now monitor the development of the active
set scheme. We can identify zig-zagging by comparing the current
active set with the active set of the second- and third-last iteration.
This very simple approach helps stabilizing the contact algorithm!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5476 060e22d8-9d1c-0410-868a-e3a950d63d95