clean up of downwinding. Name of class and file 'DownwindMatrix' is a little misleading as the thing itself is not a matrix but rather a special type of permutation operator that can permute matrices and vectors (and also can permute vectors back to original).
Note that this class does actually NOT do a global parallel downwinding (too much overhead and no use) but rather a processor local downwinding only. The reason is that we don't do true GaussSeidel in parallel anyway but rather Additive-Schwarz-GaussSeidel and so the interprocessor ordering of equations has no influence on convergence (Additive Schwarz IS overlapping block-Jacobi). No matter how processor blocks are ordered, the result will allways be the same.

Note that downwinding - of course - is NOT for structural problems, because it is a) useless there and b) combined with ML the nullspace is not accordingly permuted and does not match the permuted system anymore. Deterioration of convergence can be expected if applied.


In case of fluid/condif problems the nullspace is a dull constant in every direction (including pressure) and need not be resorted as the permutation will result in an identical nullspace.

Downwinding is not supported in SIMPLE preconditioners (will follow next)

new solver blocks include DWINDTAU to control downwinding and include preconditioners DownwindGaussSeidel and GaussSeidel, 
smoothers for ML now include SGS (Symm.GaussSeidel), GS (GaussSeidel) and DGS (Downwinding GaussSeidel)

This submit is presented to you by FEIBI

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5531 060e22d8-9d1c-0410-868a-e3a950d63d95