STRESS OUTPUT (Part whatsoever)
Enabled output of principal stresses (Cauchy and 2PK) and strains (GL)
as well as principal directions in ensight/paraview. 
When calling the ensight-filter, just type 

--stresstype=c123 (or nd123 or c123_nd123) 

for center, nodal or center and nodal principal stresses/directions.
In order to display principal directions, just use a glyph (note: for
element center principal directions, you have to use the filter 
'CenterCells' in advance!).


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5672 060e22d8-9d1c-0410-868a-e3a950d63d95