added 3d Boundary Element without any physics called Bele3 (ok, we can work on the name ;-))
added Boundary discretization (not working yet, just compiling)

together, they form a moving boundary that can carry DOFs, but do not necessarily need physics. Added also adapters for a general moving boundary and a derived XFluidFSI Boundary that can be used in a FSI Dirichlet Neumann coupling. The XFluidFSI boundary discretization will be generated once in the beginning (to be implemented), just as the ALE field in the ALE FSI formulation.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5679 060e22d8-9d1c-0410-868a-e3a950d63d95