Removal of matrix-free Newton from baci
To salve my conscience, I removed (hopefully) all matrix-free Newton
related stuff from baci. This is an ongoing construction site but
currently I am not really working on it. It will not be checked in
until the current bugs are fixed and the method is validated. Therewith
no one can accidentally use this feature and the mess in our 
strugenalpha.cpp is reduced a bit ;-)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5732 060e22d8-9d1c-0410-868a-e3a950d63d95