Introduced namespace INPUTPARAMS
This namespace should one day contain all the enums now still
defined in the various old c-headers (enums.h, dynamic.h,...)
The enums are used to represent the different values of string
input variables within baci. The idea behind this new namespace
is twofold: first, of course, the old c-headers should go away
one day and we have to establish an alternative for them.
Second, it makes sense to define the valid parameters on a
global level and not within the different types of problems.
Of course, then drt_validparameters.H has to be included
wherever we want to work with the enums!

Started the namespace with 2 enums concerning contact methods
(contact type and friction type).

If you have questions, do not hesitate to ask Uli or me!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5796 060e22d8-9d1c-0410-868a-e3a950d63d95