Major binary output cleanup
+ finally separate ccarat binary io and baci io
+ increase output flexibility (any number of control files...)
+ remove some global variables
+ quite some cleanup along the way...

Note: For the post filters to work the discretication name is bound to
be one of the known field names. We might work on that it this causes
trouble.

Note: Some information inside the control files changed. You will have
to update your post filters as well.

This change touches a lot of things. It should work fine. Partially
tested. But you never know...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5872 060e22d8-9d1c-0410-868a-e3a950d63d95