add a DebugWriter to FSI that allows to write (and visualize) unconverged interface results
This is meant as a permanent (optional) debug feature as we oftentimes have a need to see
what really happens at the FSI interface. Of course it is switched off by default.

Internally the DebugWriter creates a boundary discretization (thanks to Axel) that is
written in each time step. To enable the output of such boundary discretizations the 
problem type "none" is introduced. This is not a problem type baci can solve, however
the ensight filter has been enhanced to understand such problem types. It assumes to
find one discretization and writes whatever vectors it find in a generic way.

Note: The current implementation is 3d only as there are no 1d boundary elements...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5892 060e22d8-9d1c-0410-868a-e3a950d63d95