Move utility files from FSI to ADAPTER
The idea is that we only allow one way access

	FSI -> ADAPTER -> (single field solvers)

and thus all methods that are needed by the adapters belong to that namespace.
This way we get all the glue code in ADAPTER and end up with some slim coupling
methods in FSI. Not such a bad thing, really.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5893 060e22d8-9d1c-0410-868a-e3a950d63d95