add generic preconditioning operator to be used with MFSI
The block preconditioner in MFSI cannot use Krylov solvers for block preconditioners
as approximated Krylov methods are no linear operators and the GMRES that solves
the whole MFSI system demands a linear operator as preconditioner. Thus we just
want to apply a preconditioner on each block. So here is a new generic preconditioner
class that reads the solver parameters and builds an appropriate internal preconditioner.
Direct solvers can be used as well. These will result in the usual solver calls.

This rocks! Block solver times go down with just a slight increase of the number of
required iterations. Now MFSI is competitive with Aitken relaxed Dirichlet-Neumann
FSI.

This commit has been build by applying a knowledge transfer operator to Michael Gee.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5994 060e22d8-9d1c-0410-868a-e3a950d63d95