created method that redistributes a discretization based on node row and col maps
-> already applied it to many places, which use such functionality.

With the help of this method, the surface discretization that intersects the XFEM fluid domain is now ghosted such that all surface elements are located redundant on all processors. This way, the parallel FSI coupling is fixed for the whole computation and no expensive setup is needed for each FSI iteration step.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6032 060e22d8-9d1c-0410-868a-e3a950d63d95