use interface split block matrix for fluid field in MFSI problems
This speeds up FSI::MonolithicOverlap::SetupSystemMatrix() quite a lot.
Sadly, however, assembling to a splitted matrix is more expansive than
assembling to a plain matrix so we do not gain that much.

(There is room for improvements in block matrix assembling.)

An interessting effect is that aztec is faster when working with
block matrices. The results are the same. So this is propably due
to internal caching since we use one and the same matrix objects
when working with block matrices. On the other hand we create new
matrices for each linear system if we split. Still it's scarry.

Another advantage of block matrices here is the (assumed) reduced
memory consumption, since we do not have to split and do not need to
keep the same matrix in memory twice. (Preconditioners being equal
in both cases.)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6035 060e22d8-9d1c-0410-868a-e3a950d63d95