Launch of Beta-Version of pre_exodus. This program is meant to replace the former student 'exofilter'. Therefore the main purpose is to create a .dat file from meshing programs other than gid. However, it has a couple of improvements compared to the student-version.
First, it links to existing baci-methods, such as for example the validation methods. Therefore, any .dat-file created is automatically validated with the currently available tests (unfortunately, there are some missing, e.g. for materials or element parameters which hopefully come in the future). The validation can be perfomred also on an existing .dat-file.

Also, it provides data structure for handling and editing meshes. Any adaptation could be also done here.

Secondly, we tried to ease the process as much as possible. For example, any names on geometry or mesh entities are handed through until the .dat file. E.g. names from segmentation arrive automatically in the dat-file. The assignment of boundary conditions and element specifications has changed significantly. We refer to the new tutorial on a 3d fsi driven cavity example. Hint: Use pre_exodus --exo=yourexofile.exo to create default "head"- and "bc"-files.

The code is tested with the mentioned 3d fsi driven cavity and Georg's huge meshes. We are looking forward for any comments, suggestions, improvements, whatsoever and hope that this will simplify our life ;-)

Georg & Thomas & Moritz

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6057 060e22d8-9d1c-0410-868a-e3a950d63d95