add UnComplete() method to sparse operator
The MFSI interface matrices of structure and fluid do not have to be on the same graph.
Thus we might have to deal with additional entries when we add the fluid part to the
structural matrix. The structural matrix is already filled, thus we unfill it assemble
and fill again. This is simple but time consuming.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6105 060e22d8-9d1c-0410-868a-e3a950d63d95