Added a method PreEvaluate to the base class ElementRegister.
this method is used in the following way:

In Discretization::Evaluate there's now a new loop looping all present element register objects and calling their PreEvaluate method BEFORE the loop over column elements evaluating them.

If an instance of the derived element specific register class does not implement the method PreEvaluate the base class dummy is called that does nothing

the parameter list as well as system matrices and vectors and the discretization are passed to PreEvaluate. This gives the element specific derived register class the chance
to do stuff BEFORE the actual element evaluation is called (having the current configuration such as displacements etc present in the discretization)

In my case, the puso-tet's nodal integration over Voronoi cells is done by the puso-tet's register class.

The actual call to the puso tet itself (that follows in the usual element evaluate loop) only does the stabilization.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6141 060e22d8-9d1c-0410-868a-e3a950d63d95