add new cool matrix transformation to MFSI
MFSI builds a block system out of three independant systems. At the FSI interface the
blocks overlap. That is the structural interface displacements are taken for the unknowns
and both fluid and ale unknowns at the interface are expressed in terms of structural
displacements. So the matrix setup has to move blocks from fluid or ale maps to the
structural map. Such operations are assumed to occure more often in block systems.
Thus a general approach to such operations is assumed here. 

The refactoring of these operations has been motivated by the attempt to include
fluid linearizations with respect to the mesh motion. The outcome is a huge code
cleanup and generalization. (It looks like a simplification to me, but your mileage 
may vary.)

This is work in progress.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6253 060e22d8-9d1c-0410-868a-e3a950d63d95