further work at FSI specific matrix transformation, including support for mesh movement terms
The interesting point with these (partially finished) mesh movement terms is the reduced number
of iterations inside the linear system solver. This was not asked for as the linear iterations
are rather cheap in the current setting. The number of nonlinear (Newton) iterations did not
decrease in the pressure wave example. It was low already...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6268 060e22d8-9d1c-0410-868a-e3a950d63d95