Implemented computation of deformation gradient consistent to modified (enhanced or ANS) strains for solid elements.
As our anti-locking technique is based on E we do not provide a locking-free F at element level (ANS for sosh8 *only* works on E). In the case of Cauchy-stress evaluation, the mapping of the *locking-free* PK2-stresses is up to now based on a disp-based F. At the (high!) cost of two polar decompositions one can gain the consistent F. Just for stress mapping I find this not worth the cost and I have tried two examples with large deformations and strains, respectively, where I have hardly noticed the difference. Therefore I disconnected it.

However, if anyone likes to work on a consistent F and is willing to pay the price this code snippet should do the job.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6290 060e22d8-9d1c-0410-868a-e3a950d63d95