added Neumann-type stabilization term at outflow boundary
Note:

In case of turbulent flows, the use of a conventional "do-nothing" boundary condition at Neumann outflow boundaries may lead to simulation breakdown due to reverse flow
through that outflow boundary. Such (local) flow reversal is not unlikely because of vortices "travelling" out of the flow domain. I took much pleasure in observing repeated simulation breakdowns for flow past a square-section cylinder, where such a Neumann outflow boundary is required. Usually, for turbulent flows, an outflow boundary condition is supposed to "avoid any harm" to the "inner" flow - nothing more but also nothing less. A "do-nothing" boundary condition may serve this purpose, for instance, for laminar flows where such flow reversal is not to be expected.

The present stabilization term (proposed in Bazilevs et al. (2008)) supplements the presently implemented form of the Neumann boundary condition towards a "total-momentum" Neumann boundary condition if necessary. Since the convective term is formulated in non-conservative form and, furthermore, not integrated by parts in our current implementation, we (implicitly) employ a standard form of the Neumann boundary condition, that is, convective stresses are not part of that boundary condition. The present stabilization term contains the convective stresses and is activated at a Neumann boundary as soon as a negative velocity with respect to the outward-pointing normal vector to the Neumann boundary surface is detected. In contrast to the standard Neumann boundary condition in BACI, which is evaluated at the beginning of each time step, this NONLINEAR term needs to be evaluated at the beginning of each ITERATION step.

The following has to be done for using this outflow stabilization term which is strongly recommended when trying to simulate such problem types and expecting some reasonable results sometime somehow:

- The Neumann outflow boundary has to be declared via the respective condition. It is not sufficient to implicitly assume a "do-nothing" boundary condition, as often done. All flags and values of that condition may, of course, be set to zero.

- In the FLUID DYNAMICS/STABILIZATION, another flag OUTFLOW_STAB will be included (subsequent commit), which features the default option "no_outstab", but enables the addition of that stabilization term by setting it to "yes_outstab".

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6349 060e22d8-9d1c-0410-868a-e3a950d63d95