Added knotvectors to NURBS namespace (clean up)
----------

Included control points (which are derived from nodes with one additional
double value, the weight of the control point)

control points are added to the Factory of ParObject 
they implement nodes and hence ParObject, so they have a unique ParObjectId

----------

Derived a nurbs discretisation from the standard discretisation. 
There's nothing special in it except for the knot vector which is
added to the discretisation.
The nurbs discretisation manages control points instead of nodes, but
that doesn't change too much





git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6354 060e22d8-9d1c-0410-868a-e3a950d63d95