a first version of the single patch 2d gen-alpha fluid nurbs element (nurbs4 and nurbs9)
A SHAPEFCT parameter specifies whether we allocate a normal discretisation
or an extended nurbs version with additional knot vectors.
Input is read in the usual way, just added new methods to read knot vectors
and extended coord reader to be able to read control points instead of nodes.
Control points are derived from node and provide the weights for the nurbs
shapefunction evaluation. They are managed by the discretisation in the usual
way. 
The elements allocated in the nurbs case are fluid2nurbs, which means that
they have an overloaded Shape() call. All the line stuff is missing at the
moment (so you will get the original fluid2 methods, which do not make to much
sense in that case), but in the longrun the nurbs element will have to replace
the original lines by nurbs curves (but I haven't done that yet)

The shape function evaluation in genalpha is now done depending on the spatial
approximation type, i.e. nurbs like or normal. I had to add the local
knotvector on the element as an additional input argument (it will be empty
for the normal element)

The new postprocessing filter will follow soon (mind that we have to evaluate
shape functions during postprocessing to get function values on visualisation
points and the coordinates of the visualisation points themselves)

Input is genenerated 'manually/script based' up to now. I'm able to do order
elevation and knot insertion, but getting design lines etc for boundary
conditions on more complex geometries and so on is still far from what I
need...  
Maybe it will be possible to get nurbs geometry information from gid and have
a better, more convenient preprocessing filter.

Another major issue for the nurbs thing are periodic boundary condition
related things, i.e.
o periodic boundary conditions will require different knotvectors and the
  coupling of more than one control points (for the second order case it will
  have to couple always two pairs of points, for example 1+10 and 2+11)
o the same node coupling mechanism used for the pbc's will be necessary to
  treat problems with several control points located in the same spatial
  position (i.e. the corner of the famous plate with a hole shouldn't get two
  independent sets of degrees of freedom) 


collateral damage:

1) I had to add an additional input parameter to the element Factory to be able 
   to decide whether I want to add a nurbsX fluid2 element or a quadX fluid2
   element to my discretisation.
   That means I added dummy input parameters to the function calls in
   src/drt_constraint/constraint_manager.cpp
   src/drt_elch/elch_create_condif.cpp
   src/drt_lib/drt_condition_utils.cpp

2) The output is modified. I'm writing a flag indicating a nurbs or polynomial
   discretisation, additionally I write knot vectors (I need them for post 
   processing)
   These implied changes to io_something and fsi_debugwriter

If there are any problems with these changes please do not hesitate to contact
me.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6376 060e22d8-9d1c-0410-868a-e3a950d63d95