Prodly joining the BACI community:
This is the introduction of the combustion module!

A new namespace "COMBUST" and a probelmtype "Combustion" is introduced.
A new folder "src/drt_combust" contains combustion specific source files.

The combustion algorithm is derived from the ELCH algorithm in the context of the adapter classes. In fact, so far COMBUST is (almost) only a copy of ELCH. A 2D-ELCH-test case runs ( Haha, ein Elch-Test! Wir erinnern uns an die A-Klasse und EPS :-) )

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6499 060e22d8-9d1c-0410-868a-e3a950d63d95