1. storage of sum of incremental jumps per time step in slave nodes
Like the lagrange multipliers also the vector of the sum of the incremental jumps per time step (d_n+1 - d_n) is stored into CNode class (not only into a global vector) 

2. initialization of maps for slippy nodes / tresca friction 

A subset from the active set (nodes in contact) is the set of slippy nodes (sliding nodes). The necessary maps are initialized.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6503 060e22d8-9d1c-0410-868a-e3a950d63d95