- fixed restart of multi-scale problems - microscale solver is now read from corresponding micro-inputfile (no hard-coding anymore)   -> for this purpose, an ActivateSolver() function was introduced - in prinicple enabled usage of modified Newton on macroscale (saving a lot of homogenization time on the microscale   at the cost of additional iterations. I am still trying to find out which alternative pays off in the end...)
_Note_: 
it is currently not possible to use modified Newton at least in combination with Solidh8 (and I assume it is 
the same for the other elements except ptet). An update of calc_struct_nlnstiff is necessary and some if-statements
(currently "coupling" calculation of forces and stiffness matrix) need to be updated. 
Anyhow, even if we would change this, we still have a problem with the update of EAS parameters. Therefore I am using 
modified Newton only without EAS in my working copy, but I won't commit this hack ;-)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6571 060e22d8-9d1c-0410-868a-e3a950d63d95