Added a new nonlinear iteration: line search Newton. the implementation follows CT Kelleys book. It does a full Newton step but in case residual reduction fails (residual grows) it scales down the step length to a second guess of 0.5*disi.
If this step fails as well three points along the direction disi exist and a quadratic polynomial for the residual is minimized and an optimal step length is chosen.

The feature can be adressed from the input file by choosing 
NLNSOL lsnewton 
in the structural dynamics block.

The feature currently does not support constraint, robin or other specialties, these have to be introduced if necessary.

The feature is not yet tested for EAS elements yet but trouble can be expected from the theory as EAs currently does not allow to repeat a step once it is taken.

The implementation lives in strugenalpha_lsnewton. To shorten strugenalpha.cpp I moved FullNewton() there as well (and also to make sure that people do consistent changes to all newton methods and do not forget the LineSearchNewton routine)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6613 060e22d8-9d1c-0410-868a-e3a950d63d95