Restructuring of constraint handling:
1st part:
Introduction of a new constraint class. ConstrManager manages, as the name might suggest, instances of this new class. The Manager class looks much better now, loosing a lot of redundant code. Multi Point Constraints will be changed next.

So far the changes are rather cosmetically, but there will be changes in the element-to- ConstrManager-communication. So far elements write some of their results into the parameterlist...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6660 060e22d8-9d1c-0410-868a-e3a950d63d95