Restructuring of constraint handling:
4th part:
Elements do not write the results into the parameter list any more. Instead, when asked to evaluate a constraint boundary condition, the third element vector is used. Hence the assemble routines can be used. Within the constraint manager we still need to communicate the results between the processors.

Results are identical, but this implementation seems to be slightly faster.

All of the nightly test cases with constraints should be passed again.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6681 060e22d8-9d1c-0410-868a-e3a950d63d95