1. part of major revision of the intersection method finished: -> all of my old test cases are running again -> all the nightly test cases are still running -> several more complex problems are running now 	(such as e.g. 2.5D cases with cutter surfaces lying exactly in xfem surfaces, ...)
=> xfem elements, which are only intersected on their surfaces, 
   are now tetrahedralized as well (this case was excluded sofar, 
   but it is necessary for proper force interpolation )	,
   extension of tetgen data structure filling to cope with this problem,
=> code extended to make it less sensitive regarding further changes
=> more documentation


PLEASE NOTE: 
The intersection method is highly sensitive to any changes,
since it is impossible to test all special cases with a limited number of nightly test cases.
So, if somebody wants to change something in one of the files, which are called
"intersection...", he/she is welcome to discuss the extensions. 
Any unannounced changes will be REMOVED immediately.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6687 060e22d8-9d1c-0410-868a-e3a950d63d95