this commit is adressing the fact that the periodic boundary conditions mess up load balancing (and it's getting worse for quadratic nurbs).
What I added here is an additional call to METIS and Redistribute before I
start to set up my periodic boundary conditions.
The weighting for the METIS call is designed such that Master nodes get a
higher weight and slave nodes a lower weight (since slave nodes are moved to
the master nodes proc during construction of pbcs).  
The arising distribution is not optimal, but better then what we had before
and we get a measurable speedup, at least for the nurbs problem
 



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6691 060e22d8-9d1c-0410-868a-e3a950d63d95