amended 'calc_struct_reset_istep' operation to all structural elements. In this operation, the element-wise stored internal variables (chiefly EAS parameters) are reset to the last converged step. This is needed for time step adaptivity.
Added an empty routine MAT::ViscoNeoHooke::Reset() (while talking to Thomas Kloeppel), which does nothing, but could be helpful for future materials with a diff
erent storage techique of their internal variables subject to different evolution laws.

Subversive CommiTheme:
It will be readily admitted, that a degree conferred by an university, ought to be a pledge to the public that he who holds it possesses a certain quantity of knowledge. (Charles Babbage)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6802 060e22d8-9d1c-0410-868a-e3a950d63d95