Did a heart and kidney transplant (to stay in the biomechanical language) of the new prestressing in hex8 to wedge6
Did brief evaluation on wedge6 prestressing and get expected results.

Did not do full set of evaluation on wedge6 (including restart and poststressing) as heart and kidney proofed to be working in hex8 previously.

tet4 is still missing to be able to run an AAA with ILT FSI - tet4 is coming up next....

PS: Opposed to the medical case, hex8 still has a heart and 2 kidneys after the transplant ;-)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@6960 060e22d8-9d1c-0410-868a-e3a950d63d95