weakly imposed Dirichlet boundary conditions for the fluid2 element
the boundary conditions performs an integral over the surface:

1) factor: 2*nu
    
        /                     \
       |           s  n+af     |
     + |  v , nabla  u    * n  |
       |                       |
        \                     / boundaryele
    

2) factor: 2*nu*gamma_wd
    
        /                           \
       |        s          n+af      |
     + |   nabla  w * n , u    - u   |
       |                          b  |
        \                           / boundaryele
    

3) factor: nu*Cb/h
    
        /                \
       |        n+af      |
     + |   w , u    - u   |
       |               b  |
        \                / boundaryele
    

4) factor: 1
    
        /                             \
       |    / n+af   \       n+af      |
     - |   | u    * n | w , u    - u   |
       |    \        /              b  |
        \  |          |               / boundaryele, inflow
           +----------+
      
The gradients require data from the parent element, and they contribute to the
element matrix of the parent element (and not only to the surface element)

Thus the addidtional evaluate method in drt_discret_evaluate.cpp was necessary.

The weak Dirichlet condition is available only for quad4/line2 elements (up to
now). For other elements some things have to be updated, like the mapping of
the gausspoint of the boundary element to the parents reference domain.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7180 060e22d8-9d1c-0410-868a-e3a950d63d95