Add fixed Richardson iterations to MFSI block preconditioner fluid and structure block
MFSI depends on strong block preconditioning in order to solve the full linear system.
And since we have overlapping interface equations one block (fluid or structure) is
tainted and hard to solve. For this block it might be necessary to employ superpowers
inside the preconditioner. Enter Richardson.

Basically we solve a linear system 

	A x = b

approximatively, using just a quick preconditioner rather than a real solver. So what
we normally do is

	x = M^-1 b.

And if thats not enough we can switch on an Richardson iteration:

	x_{i+1} = x_i + omega M^-1 (b - A x_i)

with a fixed relaxation omega and a fixed number of iterations, since we are inside
a (block) preconditioner and have to behave linearly. So the convergance rate is rather
slow, but this procedure enables MFSI on problems that would otherwise be lost to
partitioned schemes...

Be careful, however, not to overuse Richardson. Interestingly it is possible to
solve the linear block systems to a high degree using this method. In this case the
global Krylov method (the actual solver of the linear block system) will require less 
iterations. But since the field coupling is managed by the Krylov iterations this is
not such a good thing. It slows down the Newton convergance, thus increasing the
number of linear block system to be solved.

So only use Richardson if the linear solver fails to converge otherwise. And yes,
you have to choose the number of iterations and the relaxation factor yourself.

Note: This commit is an implementation of advice given by Michael Gee.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7339 060e22d8-9d1c-0410-868a-e3a950d63d95