fix the restart bug that was never there ... it's just a matter of the tolerance criteria
An interesting point here is that the block preconditioner and the merged perconditioner
will produce different numerics (below the threshold). Yes, you would have expected that,
but than the difference turns out to be quite large. Especially when it comes to the
fluid pressure. The block preconditioner tends to err at the interface and you have to enforce
a propper solution in terms of the global norms. The merged preconditioner will more likely
distribute the error all over the domain.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7357 060e22d8-9d1c-0410-868a-e3a950d63d95