A new element type Combust3 is introduced to baci. It does not work yet!
For now it is only a copy of the XFluid3 element, but in the future it is supposed to combine the XFluid3 (XFEM) and Fluid3 (e.g. turbulence) functionality in a single Combust3 element combustion problems. It is used to discretize the fluid domain in a combustion problem. Condif3 will be used for the level set function domain.

in /drt_xfem/enrichment.H an additional header file is included

Both the Condif3 and Combust3 elements are associated with the D_FLUID3 defines flag. This was changed in Makefile.objects accordingly.
-> D_FLUID3 defines flag is now required for condif test cases!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7400 060e22d8-9d1c-0410-868a-e3a950d63d95