Add better (but experimental) element error handling.
The idea is to gather all element dserrors in one element call first and exit only
after the element call if there was an error. All errors will be listed in the
error file. This allows to find all broken elements in just one call, for example.

C++ exceptions are used here. 

This is framework internal. Nothing changes for element or algorithm code.

This is hidden by the THROWELEMENTERRORS flag. No change by default.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7525 060e22d8-9d1c-0410-868a-e3a950d63d95