fix fsi-*-gena buildbot failure
Now this is interessting. The cause of this sudden failure was a large drop in ML convergance
on the structural equations due to additional zeros on Dirichlet rows. Obviously we have to
respect Dirichlet rows in out matrix addition, too. This introduces Dirichlet details to an otherwise
algebraic code, but the code runs much smoother this way.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7727 060e22d8-9d1c-0410-868a-e3a950d63d95