Towards a direct solution of the constrained linear system II:
For small problems when direct solvers can be applied to solve the linear system, the zero enties on the diagonal (from the lagrange multiplier formulation) do not influence the solver's performance. Hence the input parameter UZAWAALGO is now used to switch between the usage of the Uzawa algorithm ('iterative') or the solution of the whole system with zero diagonal entries in one solver step ('direct'). 

- renamed class UzawaSolver to ConstraintSolver

- implemented a direct solve routine within this class, dealing with the whole map business

- changed input parameters as described above

- monitor evaluation consistent with constraint evaluation

For relatively small examples with only a few constraints both alternatives show comparable performance, the influence on the results are negligible.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7758 060e22d8-9d1c-0410-868a-e3a950d63d95