- proper export from overlapping to unique vector (thanks to Thomas K.)
- fixed hanging of code during output (stopped nightly parallel test last night)

%%%%%%%
Lesson: never put an LINALG::Export(...) like this:

if (myrank == 0)
{
   LINALG::Export(a,b);
   ... do, e.g. printing of vector a
}

instead do it this way:

LINALG::Export(a,b);
if (myrank == 0)
{
   ... do, e.g. printing of vector a
}
   
The Export function has to be executed on ALL processors, even if you need a result only, for instance for output, on proc 0.
Otherwise, the Export function on proc 0 is waiting forever for his colleagues on the other procs. Much head scratching was the result, so this commit shall be my personal reminder. Of course, dead simple, if you know it.
You are now allowed to laugh... ;-)



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7798 060e22d8-9d1c-0410-868a-e3a950d63d95