For incompressible flow, we get back to a non-density-weighted (or kinematic(!) with respect to viscosity and pressure) formulation by default.
Although an initial biomechanical simulation using a density orders of magnitude different from unit value by Christiane (in this case 3 orders of magnitude) did not reveal observable negative influence on solver performance, Uli conducted another biomechanical simulation, also featuring a density 3 orders of magnitude different from unit value, where a substantial negative influence on solver performance could be diagnosed. Thus, with a view on such applications, this switch back to a non-density-weighted incompressible formulation is justified, although residuals will have to be artficially weighted by density again from now on, unfortunately.

For the low-Mach-number solver, of course, nothing changes. Since density ratios are expected to be significantly less than one order of magnitude for cold-flow applications (and even for combustion applications within a range of 5-8 at maximum), potential negative influence on solver performance will be within acceptable limits for such applications, hopefully. However, the future will tell us about a potential need for some artificial scaling in this case, too.

In terms of implementation, one has to be aware of the fact that there now exist both "vede"-vectors containing node-based density values, among other things, and a (constant) scalar value "density_" which have to be distinguished. For incompressible flow, the entries in the "vede"-vectors are set to 1.0 such that the respective density weighting on the element level is of order 1.0, and the scalar value "density_" is set to the value provided in the input file, with which the subsequent artfificial weighting of the residual is performed. For low-Mach-number flow, the entries in the "vede"-vectors are set to values provided by the low-Mach-number superior steering routine such that the weighting on the element level is appropriate, and the scalar value "density_" is set to 1.0 such that the (in this case unnecessary) subsequent artificial weighting of the residual does not alter already existing values.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7804 060e22d8-9d1c-0410-868a-e3a950d63d95